requirements.txt
altgraph==0.17.3
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.8
async-generator==1.10
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
cairocffi==1.4.0
CairoSVG==2.5.2
certifi==2022.6.15
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.1.1
chromedriver==2.24.1
click==8.3.1
colorama==0.4.5
contourpy==1.0.5
cssselect2==0.7.0
customtkinter==4.6.3
cycler==0.11.0
darkdetect==0.7.1
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.6
distro==1.9.0
entrypoints==0.4
exceptiongroup==1.3.1
executing==1.0.0
fastapi==0.128.0
fastjsonschema==2.16.1
filelock==3.8.0
fonttools==4.37.4
future==0.18.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
iconify==0.0.103
idna==2.10
ipykernel==6.15.2
ipython==8.5.0
ipython-genutils==0.2.0
jedi==0.18.1
Jinja2==3.1.2
jiter==0.12.0
jsonschema==4.15.0
jupyter-core==4.11.1
jupyter_client==7.3.5
jupyterlab-pygments==0.2.2
kids.cache==0.0.7
kiwisolver==1.4.4
libsass==0.21.0
lxml==4.9.1
MarkupSafe==2.1.1
matplotlib==3.6.0
matplotlib-inline==0.1.6
mistune==2.0.4
mock==4.0.3
nbclient==0.6.7
nbconvert==7.0.0
nbformat==5.4.0
nest-asyncio==1.5.5
notebook==6.4.12
numpy==1.23.3
openai==2.14.0
outcome==1.2.0
packaging==21.3
pandocfilters==1.5.0
parso==0.8.3
pefile==2022.5.30
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==2.5.2
prometheus-client==0.14.1
prompt-toolkit==3.0.31
psutil==5.9.2
pure-eval==0.2.2
pycparser==2.21
pydantic==2.12.5
pydantic_core==2.41.5
pygame==2.1.2
Pygments==2.13.0
pyinstaller==5.4.1
pyinstaller-hooks-contrib==2022.10
pyparsing==3.0.9
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
pyrsistent==0.18.1
PySide2==5.15.2.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pywin32==304
pywin32-ctypes==0.2.0
pywinpty==2.0.7
pyzmq==23.2.1
QT-PyQt-PySide-Custom-Widgets==0.6.3
qtsass==0.3.2
requests==2.23.0
selenium==4.4.3
Send2Trash==1.8.0
shiboken2==5.15.2.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
stack-data==0.5.0
starlette==0.50.0
terminado==0.15.0
tinycss2==1.1.1
tornado==6.2
tqdm==4.64.1
traitlets==5.3.0
trio==0.21.0
trio-websocket==0.9.2
typing-inspection==0.4.2
typing_extensions==4.15.0
undetected-chromedriver==3.1.5
urllib3==1.25.11
uvicorn==0.40.0
virtualenv==20.16.5
wcwidth==0.2.5
webdriver-manager==3.8.3
webencodings==0.5.1
websockets==10.3
wsproto==1.2.0
